+++ src/modules/system/components/CAnalyseRows/CClassifyAnalyse/CClassifyAssigneesAnalyse/CClassifyAssigneesAnalyse.jsx
@@ -5,6 +5,50 @@ import CClassifyAssigneesAnalyseBar from './CClassifyAssigneesAnalyseBar';
 import CClassifyAssigneesAnalysePie from './CClassifyAssigneesAnalysePie';
 
 const CClassifyAssigneesAnalyse = () => {
+  const data = [
+    {
+      value: 18,
+      type: 'Proyang',
+      name: 'ProyangOpen',
+    },
+    {
+      value: 80,
+      type: 'Proyang',
+      name: 'ProyangClose',
+    },
+    {
+      value: 28,
+      type: 'AyaseEUmi',
+      name: 'AyaseEUmiOpen',
+    },
+    {
+      value: 55,
+      type: 'AyaseEUmi',
+      name: 'AyaseEUmiClose',
+    },
+    {
+      value: 39,
+      type: 'airoucat',
+      name: 'airoucatOpen',
+    },
+    {
+      value: 55,
+      type: 'airoucat',
+      name: 'airoucatClose',
+    },
+    {
+      value: 81,
+      type: 'Shanxiaolin',
+      name: 'ShanxiaolinOpen',
+    },
+    {
+      value: 36,
+      type: 'Shanxiaolin',
+      name: 'ShanxiaolinClose',
+    },
+  ];
+
+
   const [tabs, setState] = useState('bar');
   
   const tabList = [
@@ -18,9 +62,11 @@ const CClassifyAssigneesAnalyse = () => {
     },
   ];
   
+
+
   const contentList = {
-    bar: <CClassifyAssigneesAnalyseBar />,
-    pie: <CClassifyAssigneesAnalysePie />,
+    bar: <CClassifyAssigneesAnalyseBar data={data} />,
+    pie: <CClassifyAssigneesAnalysePie data={data} />,
   };
 
   return (
