+++ src/modules/system/components/CAnalyseRows/CClassifyAnalyse/CClassifyAssigneesAnalyse/CClassifyAssigneesAnalyseBar.jsx
@@ -2,28 +2,28 @@ import React from 'react';
 import { Chart, Geom, Axis, Tooltip, Legend, } from 'bizcharts';
 import DataSet from '@antv/data-set';
 
-const CClassifyAssigneesAnalyseBar = () => {
-  const lablesData = [
-    {
-      name: 'Open',
-      Proyang: 18,
-      AyaseEUmi: 28,
-      airoucat: 39,
-      Shanxiaolin: 81,
-    },
-    {
-      name: 'Close',
-      Proyang: 80,
-      AyaseEUmi: 55,
-      airoucat: 55,
-      Shanxiaolin: 36,
-    },
-  ];
-  const lablesDataSet = new DataSet();
-  const lablesDataValue = lablesDataSet.createView().source(lablesData);
-  lablesDataValue.transform({
+const CClassifyAssigneesAnalyseBar = props => {
+  const data = props.data;
+  var ldata = [];
+  var odata = {};
+  var cdata = {};
+  odata.name = 'Open';
+  cdata.name = 'Close';
+  for (var i = 0; i < data.length; i += 2) {
+    odata[data[i].type] = data[i].value;
+    cdata[data[i].type] = data[i+1].value;
+  }
+  ldata.push(odata);
+  ldata.push(cdata);
+  var fields = [];
+  for (var key in ldata[0]){
+    if(key!='name')fields.push(key);
+  }
+  const ds = new DataSet();
+  const dv = ds.createView().source(ldata);
+  dv.transform({
     type: 'fold',
-    fields: ['Proyang', 'AyaseEUmi', 'airoucat', 'Shanxiaolin'],
+    fields: fields,
     // 展开字段集
     key: '成员',
     // key字段
@@ -32,7 +32,7 @@ const CClassifyAssigneesAnalyseBar = () => {
 
   return (
     <div className="barLabels">
-      <Chart height={405} data={lablesDataValue} forceFit>
+      <Chart height={405} data={dv} forceFit>
         <Legend position="top" dy={-20} />
         <Axis name="成员" />
         <Axis name="Issue承担量" />
