+++ src/modules/system/components/CAnalyseRows/CClassifyAnalyse/CClassifyLabelsAnalyse/CClassifyLabelsAnalyse.jsx
@@ -7,6 +7,59 @@ import CClassifyLablesAnalysePie from './CClassifyLabelsAnalysePie';
 const CClassifyLablesAnalyse = () => {
   const [tabs, setState] = useState('bar');
   
+  const data = [
+    {
+      value: 18,
+      type: 'Bug',
+      name: 'BugOpen',
+    },
+    {
+      value: 30,
+      type: 'Bug',
+      name: 'BugClose',
+    },
+    {
+      value: 28,
+      type: 'Todo',
+      name: 'TodoOpen',
+    },
+    {
+      value: 28,
+      type: 'Todo',
+      name: 'TodoClose',
+    },
+    {
+      value: 39,
+      type: 'Doing',
+      name: 'DoingOpen',
+    },
+    {
+      value: 39,
+      type: 'Doing',
+      name: 'DoingClose',
+    },
+    {
+      value: 0,
+      type: 'Done',
+      name: 'DoneOpen',
+    },
+    {
+      value: 50,
+      type: 'Done',
+      name: 'DoneClose',
+    },
+    {
+      value: 20,
+      type: 'Problem',
+      name: 'ProblemOpen',
+    },
+    {
+      value: 20,
+      type: 'Problem',
+      name: 'ProblemClose',
+    },
+  ];
+
   const tabList = [
     {
       key: 'bar',
@@ -19,8 +72,8 @@ const CClassifyLablesAnalyse = () => {
   ];
   
   const contentList = {
-    bar: <CClassifyLablesAnalyseBar />,
-    pie: <CClassifyLablesAnalysePie />,
+    bar: <CClassifyLablesAnalyseBar data={data}/>,
+    pie: <CClassifyLablesAnalysePie data={data}/>,
   };
 
   return (
