+++ src/modules/system/components/CAnalyseRows/CClassifyAnalyse/CClassifyLabelsAnalyse/CClassifyLabelsAnalysePie.jsx
@@ -9,60 +9,9 @@ import {
 } from 'bizcharts';
 import DataSet from '@antv/data-set';
 
-const CClassifyLablesAnalysePie = () => {
+const CClassifyLablesAnalysePie = props => {
   const { DataView } = DataSet;
-  const data = [
-    {
-      value: 18,
-      type: 'Bug',
-      name: 'BugOpen',
-    },
-    {
-      value: 30,
-      type: 'Bug',
-      name: 'BugClose',
-    },
-    {
-      value: 28,
-      type: 'Todo',
-      name: 'TodoOpen',
-    },
-    {
-      value: 28,
-      type: 'Todo',
-      name: 'TodoClose',
-    },
-    {
-      value: 39,
-      type: 'Doing',
-      name: 'DoingOpen',
-    },
-    {
-      value: 39,
-      type: 'Doing',
-      name: 'DoingClose',
-    },
-    {
-      value: 0,
-      type: 'Done',
-      name: 'DoneOpen',
-    },
-    {
-      value: 50,
-      type: 'Done',
-      name: 'DoneClose',
-    },
-    {
-      value: 20,
-      type: 'Problem',
-      name: 'ProblemOpen',
-    },
-    {
-      value: 20,
-      type: 'Problem',
-      name: 'ProblemClose',
-    },
-  ];
+  const data=props.data;
   const dv = new DataView();
   dv.source(data).transform({
     type: 'percent',
